break 1 circular
