+++ src/symbol/symbol_layout.ts
@@ -15,7 +15,7 @@ import classifyRings from '../util/classify_rings';
 import EXTENT from '../data/extent';
 import SymbolBucket from '../data/bucket/symbol_bucket';
 import EvaluationParameters from '../style/evaluation_parameters';
-import {SIZE_PACK_FACTOR} from './symbol_size';
+import {SIZE_PACK_FACTOR, MAX_PACKED_SIZE, MAX_GLYPH_ICON_SIZE} from './symbol_size';
 import ONE_EM from './one_em';
 import type {CanonicalTileID} from '../source/tile_id';
 import type {Shaping, PositionedIcon, TextJustify} from './shaping';
@@ -486,10 +486,6 @@ function addFeature(bucket: SymbolBucket,
         }
     }
 }
-
-const MAX_GLYPH_ICON_SIZE = 255;
-const MAX_PACKED_SIZE = MAX_GLYPH_ICON_SIZE * SIZE_PACK_FACTOR;
-export {MAX_PACKED_SIZE};
 
 function addTextVertices(bucket: SymbolBucket,
     anchor: Point,
